backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1

# AI/ML
anthropic>=0.18.1
openai>=1.12.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Task Queue
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Utilities
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.19.0
